requirements.txt
requests==2.31.0
pytest==7.4.0
allure-pytest==2.13.2
pyyaml==6.0
jsonpath==0.82
pymysql==1.1.0
loguru==0.7.2
